feat: moved dnd toy model to dnd-example.tsx, and made tsx export Droppable and Draggable components
